Clean <Grid /> test file: The constants used in the tests were placed into files
