+++ test/Grid.spec.js
@@ -1,7 +1,11 @@
 import Adapter from 'enzyme-adapter-react-16';
 import Axios from 'axios';
+import bodyRenderer from './utils/bodyRenderer.js';
+import columnStructure from './utils/columnStructure.js';
+import data from './utils/data.js';
 import Enzyme from 'enzyme';
 import fakeColumnStructure from './utils/fakeColumnStructure.js';
+import footerRenderer from './utils/footerRenderer.js';
 import Grid from '../src/Grid/Grid';
 import MockAdapter from 'axios-mock-adapter';
 import orders from './utils/orders.json';
@@ -30,129 +34,8 @@ describe('<Grid />', () => {
     mount = createMount();
     shallow = createShallow({ dive: true });
   });
-  
-  const columns = [
-    {
-      'Label': 'Order ID',
-      'Name': 'OrderID',
-      'Sortable': true,
-      'SortOrder': 1,
-      'SortDirection': 'Ascending',
-      'IsKey': true,
-      'Searchable': false,
-      'Visible': true,
-      'Filter': { 
-        Argument: [],
-        HasFilter: true,
-        Name: 'OrderID',
-        Operator: 'None',
-        OptionsUrl: null,
-        Text: null
-      },
-      'DataType': 'numeric',
-      'Aggregate': 'None'
-    },
-    {
-      'Label': 'Customer Name',
-      'Name': 'CustomerName',
-      'Sortable': true,
-      'SortOrder': -1,
-      'SortDirection': 'None',
-      'IsKey': false,
-      'Searchable': true,
-      'Visible': true,
-      'Filter': { 
-        Argument: [],
-        HasFilter: true,
-        Name: 'CustomerName',
-        Operator: 'None',
-        OptionsUrl: null,
-        Text: null
-      },
-      'DataType': 'string',
-      'Aggregate': 'None'
-    },
-    {
-      'Label': 'Shipped Date',
-      'Name': 'ShippedDate',
-      'Sortable': true,
-      'SortOrder': -1,
-      'SortDirection': 'None',
-      'IsKey': false,
-      'Searchable': false,
-      'Visible': true,
-      'Filter': { 
-        Argument: [],
-        HasFilter: true,
-        Name: 'ShippedDate',
-        Operator: 'None',
-        OptionsUrl: null,
-        Text: null
-      },
-      'DataType': 'datetime',
-      'Aggregate': 'None'
-    },
-    {
-      'Label': 'Shipper City',
-      'Name': 'ShipperCity',
-      'Sortable': true,
-      'SortOrder': -1,
-      'SortDirection': 'None',
-      'IsKey': false,
-      'Searchable': false,
-      'Visible': true,
-      'Filter': { 
-        Argument: [],
-        HasFilter: true,
-        Name: 'ShipperCity',
-        Operator: 'None',
-        OptionsUrl: null,
-        Text: null
-      },
-      'DataType': 'string',
-      'Aggregate': 'None'
-    }
-  ];
-  const data = [
-    [1, 'Microsoft', '2016-03-19T19:00:00', 'Guadalajara, JAL, Mexico'], 
-    [2, 'Microsoft', '2016-04-23T10:00:00', 'Guadalajara, JAL, Mexico'], 
-    [3, 'Microsoft', '2016-12-22T08:00:00', 'Guadalajara, JAL, Mexico'], 
-    [4, 'Unosquare LLC', '2016-02-01T18:00:00', 'Los Angeles, CA, USA'], 
-    [5, 'Microsoft', '2016-11-10T18:00:00', 'Guadalajara, JAL, Mexico'], 
-    [6, 'Unosquare LLC', '2016-11-06T18:00:00', 'Los Angeles, CA, USA'], 
-    [7, 'Unosquare LLC', '2016-11-11T18:00:00', 'Leon, GTO, Mexico'], 
-    [8, 'Unosquare LLC', '2016-11-08T18:00:00', 'Portland, OR, USA'], 
-    [9, 'Vesta', '2016-11-07T18:00:00', 'Guadalajara, JAL, Mexico'], 
-    [10, 'Unosquare LLC', '2016-11-05T18:00:00', 'Portland, OR, USA']
-  ];
 
   const aggregate = { CustomerName: 500 };
-
-  const bodyRenderer = (row, index) => 
-    <TableRow hover key = { index }>
-      <TableCell padding = { 'default' }>
-        { row.OrderID}
-      </TableCell>
-      <TableCell padding = { 'default' }>
-        { row.CustomerName}
-      </TableCell>
-      <TableCell padding = { 'default' }>
-        { row.ShippedDate}
-      </TableCell>
-      <TableCell padding = { 'default' }>
-        { row.ShipperCity}
-      </TableCell>
-    </TableRow>;
-
-  const footerRenderer = aggregates => 
-    <TableFooter>
-      <TableRow>
-        <TableCell>Total: </TableCell>
-        <TableCell> { aggregates && aggregates.CustomerName } </TableCell>
-        <TableCell> ~~~ </TableCell>
-        <TableCell> ~~~ </TableCell>
-      </TableRow>
-    </TableFooter>;
 
   const grid = () => {
     if(!mountedGrid){
@@ -171,10 +54,9 @@ describe('<Grid />', () => {
     expect(wrapper).to.have.lengthOf(1);
   });
 
-
   it('should render n columns', () => {
     const columns = grid().find(TableHead).find(TableRow).find(TableCell);
-    expect(columns).to.have.lengthOf(4);
+    expect(columns).to.have.lengthOf(5);
   });
       
   it('should have 1 rows at first', () => {
@@ -200,14 +82,14 @@ describe('<Grid />', () => {
 
   describe('When custom body is defined', () => {
     it('should render the custom body', () => {
-      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columns }).reply(200, {
+      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columnStructure }).reply(200, {
         orders
       });
 
       const grid = <Grid 
-        dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns) }
+        dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure) }
         bodyRenderer = { bodyRenderer }
-        columns = { columns } />;
+        columns = { columnStructure } />;
 
       const wrapper = shallow(grid);
       const body = wrapper.find(Table).find(TableBody);
@@ -234,12 +116,12 @@ describe('<Grid />', () => {
   /** Unit tests for custom footer */
   describe('When footer has no rows', () => {
     it('should not render any row', () => {
-      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columns }).reply(200, {
+      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columnStructure }).reply(200, {
         orders
       });
     
-      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns) }
-        columns = { columns }
+      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure) }
+        columns = { columnStructure }
       />;
 
       const wrapper = shallow(grid);
@@ -251,14 +133,14 @@ describe('<Grid />', () => {
 
   describe('When footer has n rows', () => {
     beforeEach(() => {
-      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columns }).reply(200, {
+      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columnStructure }).reply(200, {
         orders
       });
     });
 
     it('should render the row with the aggregate operation', () => {
-      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns) }
-        columns = { columns }
+      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure) }
+        columns = { columnStructure }
         footerRenderer = { footerRenderer }
       />;
 
@@ -270,8 +152,8 @@ describe('<Grid />', () => {
     });
 
     it('should render the row with the bottom pager', () => {
-      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns) }
-        columns = { columns }
+      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure) }
+        columns = { columnStructure }
         showBottomPager = { true }
       />;
 
@@ -282,8 +164,8 @@ describe('<Grid />', () => {
     });
 
     it('should render the rows with the aggregate operation and the bottom pager', () => {
-      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns) }
-        columns = { columns }
+      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure) }
+        columns = { columnStructure }
         showBottomPager = { true }
         footerRenderer = { footerRenderer }
       />;
@@ -298,14 +180,14 @@ describe('<Grid />', () => {
 
   describe('When footer has showBottomPager property set as true', () => {
     beforeEach(() => {
-      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columns }).reply(200, {
+      mock.onPost('http://tubular.azurewebsites.net/api/orders/paged', { columnStructure }).reply(200, {
         orders
       });
     });
 
     it('should have a paginator', () => {
-      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns) }
-        columns = { columns }
+      const grid = <Grid dataSource = { new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure) }
+        columns = { columnStructure }
         showBottomPager = { true }
       />;
 
@@ -318,7 +200,7 @@ describe('<Grid />', () => {
 
   beforeEach(() => {
     props = {
-      dataSource: new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columns)
+      dataSource: new RemoteDataSource('http://tubular.azurewebsites.net/api/orders/paged', columnStructure)
     };
     axiosInstance = Axios.create();
     mock = new MockAdapter(axiosInstance);
