+++ test/utils/columnStructure.js
@@ -1,5 +1,8 @@
 const columns = [
   {
+    Label: 'Order ID',
+    Name: 'OrderID',
+    Sortable: true,
     SortOrder: 1,
     SortDirection: 'Descending',
     IsKey: true,
@@ -29,5 +32,4 @@ const columns = [
     DataType: 'numeric'
   }
 ];
-
 export default columns;
